+++ functions/api/image/nfts/collection/[index].tsx
@@ -2,8 +2,8 @@
 import { ImageResponse } from '@vercel/og'
 import React from 'react'
 
-import { getColor } from '../../../../../src/hooks/useColor'
 import { blocklistedCollections } from '../../../../../src/nft/utils/blocklist'
+import { getColor } from '../../../../../src/utils/getColor'
 import { CHECK_URL, WATERMARK_URL } from '../../../../constants'
 import getCollection from '../../../../utils/getCollection'
 import getFont from '../../../../utils/getFont'
