+++ functions/api/image/tokens/[[index]].tsx
@@ -2,7 +2,7 @@
 import { ImageResponse } from '@vercel/og'
 import React from 'react'
 
-import { getColor } from '../../../../src/hooks/useColor'
+import { getColor } from '../../../../src/utils/getColor'
 import { WATERMARK_URL } from '../../../constants'
 import getFont from '../../../utils/getFont'
 import getNetworkLogoUrl from '../../../utils/getNetworkLogoURL'
