+++ functions/utils/getColor.test.ts
@@ -1,5 +1,5 @@
 import { DEFAULT_COLOR } from '../../src/constants/tokenColors'
-import { getColor } from '../../src/hooks/useColor'
+import { getColor } from '../../src/utils/getColor'
 
 test('should return the average color of a black PNG image', async () => {
   const image = 'https://static.vecteezy.com/system/resources/previews/001/209/957/original/square-png.png'
